requirement.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
pigpio==1.78
pigpio-dht==0.3.6
Werkzeug==3.1.3